Fix project structure
Файлы стилей перемещены к компонентам использующих их. Компоненты страниц перемещены в отдельную папку.
