+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import Navigation from "./components/Navigation";
 import {Outlet} from "react-router-dom";
 import Footer from "./components/Footer";
-import './sass/home/sections/header.sass';
+import './pages/Home/components/Header.sass';
 
 function App() {
   return (
