+++ src/pages/Home/Home.sass
@@ -1,0 +1,2 @@
+@import components/Header, SectionTitle, components/ChooseYourTour/ChooseYourTourCards, CardCreation, components/TourHistory
+@import components/ChooseYourTour/ChooseYourTour, components/BuildYourTour, components/Feedback, components/TourPhotos, components/CTASection
