+++ src/pages/Home/components/BuildYourTour/BuildYourTour.sass
@@ -1,6 +1,6 @@
-@import buildYourTourComponent/ageAndLicenceInput
-@import buildYourTourComponent/formBuildingButtons
-@import buildYourTourComponent/inputBuilding
+@import BuildYourTour/buildYourTourComponents/ageAndLicenceInput
+@import BuildYourTour/buildYourTourComponents/formBuildingButtons
+@import BuildYourTour/buildYourTourComponents/inputBuilding
 
 
 section.build-your-tour
