+++ src/pages/Home/components/ChooseYourTour/ChooseYourTour.sass
@@ -1,4 +1,4 @@
-@import chooseYourTourCards
+@import ChooseYourTour/ChooseYourTourCards
 
 section.choose-your-tour
   +sectionInit
