--- src/sass/home/home.sass
@@ -1,2 +1,0 @@
-@import sections/header, sectionTitle, sections/chooseYourTourCards, cardCreation, sections/tourHistory
-@import sections/chooseYourTour, sections/buildYourTour, sections/feedback, sections/tour-photos, sections/cta
